+++ spec/spec/element.js
@@ -677,15 +677,6 @@ describe('Element', function() {
       expect(clone.get(0).id()).not.toBe(group.get(0).id())
       expect(clone.get(1).id()).not.toBe(group.get(1).id())
     })
-    it('inserts the clone after the cloned element', function() {
-      clone = rect.clone()
-      expect(rect.next()).toBe(clone)
-    })
-    it('inserts the clone in the specified parent', function() {
-      var g = draw.group()
-      clone = rect.clone(g)
-      expect(g.get(0)).toBe(clone)
-    })
     it('deep copies over dom data', function() {
       group.dom = {'foo':'bar'}
       rect.dom = {'foo':'baz'}